move & split conflict stuff
