+++ src/main/scala/rover/rdo/conflict/ConflictResolutionMechanism.scala
@@ -1,0 +1,7 @@
+package rover.rdo.conflict
+
+import rover.rdo.AtomicObjectState
+
+trait ConflictResolutionMechanism[A] {
+	def resolveConflict(conflictedState: ConflictedState[A]): AtomicObjectState[A]
+}
